backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
bcrypt==4.2.1
google-auth==2.37.0
python-multipart==0.0.18
slowapi==0.1.9
httpx==0.27.2
cryptography==41.0.7
python-jose[cryptography]==3.3.0
Pillow==11.0.0
firebase-admin==6.6.0
APScheduler==3.10.4
email-validator==2.2.0
